pyproject.toml
[project]
name = "debate-mas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "tyl-flora", email = "tyl021028@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "clickhouse-connect>=0.10.0",
    "duckduckgo-search>=8.1.1",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pypdf>=6.6.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "quantchdb>=0.1.11",
    "scipy>=1.17.0",
    "tabulate>=0.9.0",
]

[project.scripts]
debate-mas = "debate_mas:main"

[build-system]
requires = ["uv_build>=0.9.25,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=9.0.2",
]
